requirements.txt
Pillow
PyQt5